% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_date_features}
\alias{add_date_features}
\title{Add date-based feature columns to a data frame}
\usage{
add_date_features(df)
}
\arguments{
\item{df}{A data frame with at least a column called \code{date}}
}
\description{
Add date-based feature columns to a data frame
}
