% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cummedian}
\alias{cummedian}
\title{Naive median equivalent to \code{dplyr::cummean()}}
\usage{
cummedian(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{Remove NAs (passed to \code{median()})}
}
\description{
Naive median equivalent to \code{dplyr::cummean()}
}
