% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panning_functions.R
\name{GeneralStep}
\alias{GeneralStep}
\title{General step of panning algorithm}
\usage{
GeneralStep(y, X, Id_1s, pi = 0.5, B = 500L, d, alpha = 0.05,
  seed = 854751L, K = 10L, m = 10L, family, type = NULL, divergence,
  W = NULL, proc = 1L, C0 = 0.5, increasing = FALSE, trace = TRUE,
  ...)
}
\arguments{
\item{y, X, m, K, family, type, divergence, C0, W, increasing, trace, ...}{(see function \code{\link[panning]{CVmFold}})}

\item{Id_1s}{is the set of indices of promising variables of model of size \code{d-1}.}

\item{pi}{is the probability of selecting a predictor from \code{Id_1s}.}

\item{d, alpha, B, seed, proc}{(see function \code{\link[panning]{InitialStep}})}
}
\value{
\code{GeneralStep} returns a list with the following components (exactly the same as in
\code{\link[panning]{InitialStep}}):
\describe{
     \item{\code{Ids}}{is the set \eqn{I_d^*} of indices of predictors with prediction errors
     \code{cv.error}<= \code{q.alpha}.}
     \item{\code{Sds}}{is the set \eqn{S_d^*} of models of size \code{d} with
     prediction errors \code{cv.error}<= \code{q.alpha}.}
     \item{\code{cv.error}}{is a (\code{B} x 1) vector of CV predictions errors.}
     \item{\code{q.alpha}}{is the empirical \code{alpha}-quantile computed on \code{cv.error}.}
     \item{\code{var.mat}}{is a (\code{B}x\code{d}) matrix of indices of the explored models.}
}
The indices returned by \code{Ids} are the column number of \code{X} as it is inputed,
and not the name of the column. The indices are sorted by increasing number. Duplicates
are deleted. \code{Sds} may contain duplicates.
}
\description{
\code{GeneralStep} computes the intial step of the Panning Algorithm.
}
\details{
This function computes the m-fold Cross-validation (CV) prediction error for \code{B} models
of size \code{d}. Each of those \code{B} models are randomly constructed with the following scheme:
a predictor has a probability \code{pi} to be selected from \code{Id_1s} and a probability
1-\code{pi} from its complement; a predictor can appear at maximum once in one model (no replacement
within a model).

The \code{seed} can be fixed for reproducibility.

This function is computationnaly time consuming proportionally to the size of \code{B}.
}
\examples{
#####
# Simulate a logistic regression
n <- 50
set.seed(123)
beta <- c(1, rpois(40, lambda = 0.5))
p <- length(beta)
X <- matrix(rnorm((p-1)*n), nrow=n, ncol=(p-1))
y <- rbinom(n,1,1/(1+exp(-tcrossprod(beta, cbind(1, X)))))
#####
# Assume that Id_1s obtained from the Initial Step is (see example in \\code{\\link[panning]{InitialStep}})
Id_1s <- c(24,33)
# (can take several seconds to run)
GStep <- GeneralStep(y = y, X = X, Id_1s = c(24,33), d = 2, B = 50, family = binomial(link = "logit"),
                     type = "response", divergence = "classification", trace = FALSE)

# Run the parallelised version (4 cores)
GStep <- GeneralStep(y = y, X = X, Id_1s = c(24,33), d = 2, B = 50, family = binomial(link = "logit"),
                     type = "response", divergence = "classification", proc = 4, trace = FALSE)

}
\author{
Samuel Orso \email{Samuel.Orso@unige.ch}
}
\references{
Guerrier, S., Mili, N., Molinari, R., Orso, S., Avella-Medina, M. and Ma, Y. (2015)
A Paradigmatic Regression Algorithm for Gene Selection Problems.
\emph{submitted manuscript}. \url{http://arxiv.org/abs/1511.07662}.
}
\seealso{
\code{\link[panning]{CVmFold}}, \code{\link[panning]{InitialStep}}
}

