% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panning_functions.R
\name{InitialStep}
\alias{InitialStep}
\title{Initial step of the panning algorithm}
\usage{
InitialStep(y, X, d = 1L, alpha = 0.05, B = NULL, seed = 951L,
  m = 10L, K = 10L, family, type = NULL, divergence, W = NULL,
  proc = 1L, C0 = 0.5, increasing = FALSE, trace = TRUE, ...)
}
\arguments{
\item{y, X, m, K, family, type, divergence, C0, W, increasing, trace, ...}{(see function \code{\link[panning]{CVmFold}})}

\item{d}{the dimension of the model of interest (intercept is always included).}

\item{alpha}{the level of the quantile of the prediction errors.}

\item{B}{the number of bootstrap replicates.}

\item{seed}{the seed for the random number generator.}

\item{proc}{number of processor(s) for parallelisation.}
}
\value{
\code{InitialStep} returns a list with the following components:
\describe{
     \item{\code{Ids}}{is the set \eqn{I_d^*} of indices of predictors with prediction errors
     \code{cv.error}<= \code{q.alpha}.}
     \item{\code{Sds}}{is the set \eqn{S_d^*} of models of size \code{d} with
     prediction errors \code{cv.error}<= \code{q.alpha}.}
     \item{\code{cv.error}}{is a (\code{B} x 1) vector of CV predictions errors.}
     \item{\code{q.alpha}}{is the empirical \code{alpha}-quantile computed on \code{cv.error}.}
     \item{\code{var.mat}}{is a (\code{B}x\code{d}) matrix of indices of the explored models.}
}
The indices returned by \code{Ids} are the column number of \code{X} as it is inputed,
and not the name of the column. The indices are sorted by increasing number. Duplicates
are deleted. \code{Sds} may contain duplicates.
}
\description{
\code{InitialStep} computes the intial step of the Panning Algorithm.
}
\details{
This function computes exhaustively the m-fold Cross-validation (CV) prediction error
for all the \eqn{\left( {\begin{array}{*{20}{c}} p \\ d \end{array}} \right)}{C(p,d)} possible models of size \code{d} by calling
the \code{\link[panning]{CVmFold}} function. If \code{B=NULL} (default), then
\code{B} is set to be equal to \eqn{\left( {\begin{array}{*{20}{c}} p \\ d \end{array}} \right)}{C(p,d)}.

If \code{B} takes a positive integer value smaller than the total number of models \eqn{\left( {\begin{array}{*{20}{c}} p \\ d \end{array}} \right)}{C(p,d)},
then the function computes the CV prediction errors for \code{B} models of size \code{d} randomly selected.
In this case, it is possible to set the \code{seed} for reproducibility.

At this stage, the algorithm does not allow for interaction terms among variables.

This function is computationnaly time consuming proportionally to the size of \code{B}.
}
\examples{
\dontrun{
#####
# Simulate a logistic regression
n <- 50
set.seed(123)
beta <- c(1, rpois(40, lambda = 0.5))
p <- length(beta)
X <- matrix(rnorm((p-1)*n), nrow=n, ncol=(p-1))
y <- rbinom(n,1,1/(1+exp(-tcrossprod(beta, cbind(1, X)))))
#####

# (can take several seconds to run)
IStep <- InitialStep(y = y, X = X, family = binomial(link = "logit"), type = "response",
                     divergence = "classification", trace = FALSE)

# Run the parallelised version (4 cores)
IStep <- InitialStep(y = y, X = X, family = binomial(link = "logit"), type = "response",
                     divergence = "classification", proc = 2, trace = FALSE)
}
}
\author{
Samuel Orso \email{Samuel.Orso@unige.ch}
}
\references{
Guerrier, S., Mili, N., Molinari, R., Orso, S., Avella-Medina, M. and Ma, Y. (2015)
A Paradigmatic Regression Algorithm for Gene Selection Problems.
\emph{submitted manuscript}. \url{http://arxiv.org/abs/1511.07662}.
}
\seealso{
\code{\link[panning]{CVmFold}}, \code{\link[panning]{GeneralStep}}
}

