% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panning-package.R
\docType{package}
\name{panning-package}
\alias{panning}
\alias{panning-package}
\title{panning: Robust Median Filter Forecast}
\description{
Forecast univariate monthly timeseries using a model robust against noise and shifts at year change.
}
\author{
\strong{Maintainer}: Tim Radtke \email{tim@timradtke.net}

}
\keyword{internal}
