% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{panning}
\alias{panning}
\title{Generate forecast sample paths from an \code{rmf()} model}
\usage{
panning(object, h = 12, n_samples = 5000)
}
\arguments{
\item{object}{List returned by \code{rmf()}}

\item{h}{Forecast horizon (number of months into the future to predict)}

\item{n_samples}{Number of sample paths to be generated}
}
\description{
Generate forecast sample paths from an \code{rmf()} model
}
