% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{preprocess}
\alias{preprocess}
\title{Return differenced series prepared as training data frame}
\usage{
preprocess(y, dates)
}
\arguments{
\item{y}{A numeric vector}

\item{dates}{A date vector of equal length as \code{y}}
}
\value{
Data frame
}
\description{
Return differenced series prepared as training data frame
}
