% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf.R
\name{rmf}
\alias{rmf}
\title{Fits a Robust Median Filter model to a monthly timeseries}
\usage{
rmf(y, dates, use_seasonality = FALSE)
}
\arguments{
\item{y}{A numeric vector representing a univariate timeseries to be forecast}

\item{dates}{A date vector}

\item{use_seasonality}{logical; if \code{TRUE}, monthly dummies are added to the
forecast model}
}
\value{
A list of a fitted model and other objects
}
\description{
Fits a Robust Median Filter model to a monthly timeseries
}
