% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize.R
\name{synthesize}
\alias{synthesize}
\title{Create randomly sampled time series}
\usage{
synthesize(dates)
}
\arguments{
\item{dates}{A vector of months along which to sample}
}
\value{
A numeric vector
}
\description{
Create randomly sampled time series
}
\examples{
dates <- seq(as.Date("2016-01-01"), as.Date("2022-12-01"), by = "month")

synthesize(dates = dates)

}
