% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panning-package.r
\docType{package}
\name{panning}
\alias{panning}
\alias{panning-package}
\title{An implementation of the Panning Algorithm}
\description{
Functions to run the Panning Algorithm in R.
}
\details{
\tabular{ll}{
Package: \tab panning\cr
Type: \tab Package\cr
Version: \tab 0.2.0\cr
Date: \tab 2016-04-07\cr
License: \tab GPL 2\cr
}
}
\author{
Samuel Orso \email{Samuel.Orso@unige.ch},
Roberto Molinari \email{roberto.molinari@unige.ch},
Stephane Guerrier \email{stephane@illinois.edu},
James Balamuta \email{balamut2@illinois.edu},
Nabil Mili,
Marco Avella-Medina,
Yanyuan Ma

Stephane Guerrier \email{stephane@illinois.edu}
}

